requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.5
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1
pillow>=10.2.0
ffmpeg-python==0.2.0

